IndexedDB/idbdatabase_createObjectStore11.htm

<!DOCTYPE html>
<title>IDBDatabase.createObjectStore() - attempt to create an object store with no argument </title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://dvcs.w3.org/hg/IndexedDB/raw-file/tip/Overview.html#widl-IDBDatabase-createObjectStore-IDBObjectStore-DOMString-name-IDBObjectStoreParameters-optionalParameters">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support.js"></script>
<div id="log"></div>
<script>
var t = async_test(),
    open_rq = createdb(t);

open_rq.onupgradeneeded = function(e) {
    var db = e.target.result;

    assert_throws(new TypeError(), function() {
        db.createObjectStore();
    });

    t.done();
}
</script>
